backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
aiosqlite==0.19.0
supabase==2.3.0
